rename to electron_diffraction1d/2d.py
