+++ pyxem/signals/diffraction_simulation.py
@@ -19,7 +19,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
 
-from pyxem.signals.electron_diffraction import ElectronDiffraction2D
+from pyxem.signals.electron_diffraction2d import ElectronDiffraction2D
 
 
 class DiffractionSimulation:
