+++ pyxem/signals/electron_diffraction1d.py
