+++ pyxem/tests/test_generators/test_calibration_generator.py
@@ -22,11 +22,11 @@ import numpy as np
 from hyperspy.signals import Signal2D
 from hyperspy.roi import Line2DROI
 
-from pyxem.signals.electron_diffraction import ElectronDiffraction2D
+from pyxem.signals.electron_diffraction2d import ElectronDiffraction2D
 from pyxem.generators.calibration_generator import CalibrationGenerator
 from pyxem.utils.calibration_utils import generate_ring_pattern
 from pyxem.libraries.calibration_library import CalibrationDataLibrary
-from pyxem.signals.electron_diffraction import ElectronDiffraction2D
+from pyxem.signals.electron_diffraction2d import ElectronDiffraction2D
 
 
 @pytest.fixture
