+++ pyxem/tests/test_generators/test_subpixelrefinement_generator.py
@@ -21,7 +21,7 @@ import numpy as np
 
 from pyxem.generators.subpixelrefinement_generator import SubpixelrefinementGenerator
 from pyxem.signals.diffraction_vectors import DiffractionVectors
-from pyxem.signals.electron_diffraction import ElectronDiffraction2D
+from pyxem.signals.electron_diffraction2d import ElectronDiffraction2D
 from skimage import draw
