+++ pyxem/tests/test_generators/test_vdf_generator.py
@@ -21,7 +21,7 @@ import numpy as np
 
 from pyxem.generators.vdf_generator import VDFGenerator
 
-from pyxem.signals.electron_diffraction import ElectronDiffraction2D
+from pyxem.signals.electron_diffraction2d import ElectronDiffraction2D
 from pyxem.signals.diffraction_vectors import DiffractionVectors
 from pyxem.signals.vdf_image import VDFImage
