+++ pyxem/tests/test_library/test_calibration_library.py
@@ -22,7 +22,7 @@ import numpy as np
 from hyperspy.signals import Signal2D
 from hyperspy.roi import Line2DROI
 
-from pyxem.signals.electron_diffraction import ElectronDiffraction2D
+from pyxem.signals.electron_diffraction2d import ElectronDiffraction2D
 from pyxem.libraries.calibration_library import CalibrationDataLibrary
