+++ pyxem/tests/test_physical/test_strain_mapping.py
@@ -23,7 +23,7 @@ import numpy as np
 
 from pyxem.components.scalable_reference_pattern import ScalableReferencePattern
 from pyxem.generators.diffraction_generator import DiffractionGenerator
-from pyxem.signals.electron_diffraction import ElectronDiffraction2D
+from pyxem.signals.electron_diffraction2d import ElectronDiffraction2D
 
 
 def test_strain_mapping_affine_transform():
