+++ pyxem/tests/test_signals/test_diffraction_simulation.py
@@ -21,7 +21,7 @@ import pytest
 from pyxem.signals.diffraction_simulation import DiffractionSimulation
 from pyxem.signals.diffraction_simulation import ProfileSimulation
 from pyxem.generators.diffraction_generator import DiffractionGenerator
-from pyxem.signals.electron_diffraction import ElectronDiffraction2D
+from pyxem.signals.electron_diffraction2d import ElectronDiffraction2D
 
 
 @pytest.fixture
