+++ pyxem/tests/test_signals/test_electron_diffraction1d.py
@@ -18,40 +18,40 @@
 
 import numpy as np
 import pytest
-from pyxem.signals.diffraction_profile import ElectronDiffraction1D
+from pyxem.signals.electron_diffraction1d import ElectronDiffraction1D
 
 
 class TestSimpleHyperspy:
     # Tests functions that assign to hyperspy metadata
 
-    def test_set_experimental_parameters(self, diffraction_profile):
-        diffraction_profile.set_experimental_parameters(accelerating_voltage=3,
+    def test_set_experimental_parameters(self, electron_diffraction1d):
+        electron_diffraction1d.set_experimental_parameters(accelerating_voltage=3,
                                                         camera_length=3,
                                                         scan_rotation=1,
                                                         convergence_angle=1,
                                                         rocking_angle=1,
                                                         rocking_frequency=1,
                                                         exposure_time=1)
-        assert isinstance(diffraction_profile, ElectronDiffraction1D)
+        assert isinstance(electron_diffraction1d, ElectronDiffraction1D)
 
-    def test_set_scan_calibration(self, diffraction_profile):
-        diffraction_profile.set_scan_calibration(19)
-        assert isinstance(diffraction_profile, ElectronDiffraction1D)
+    def test_set_scan_calibration(self, electron_diffraction1d):
+        electron_diffraction1d.set_scan_calibration(19)
+        assert isinstance(electron_diffraction1d, ElectronDiffraction1D)
 
     @pytest.mark.parametrize('calibration', [1, 0.017, 0.5, ])
     def test_set_diffraction_calibration(self,
-                                         diffraction_profile,
+                                         electron_diffraction1d,
                                          calibration):
-        diffraction_profile.set_diffraction_calibration(calibration)
-        dx = diffraction_profile.axes_manager.signal_axes[0]
+        electron_diffraction1d.set_diffraction_calibration(calibration)
+        dx = electron_diffraction1d.axes_manager.signal_axes[0]
         assert dx.scale == calibration
 
 
 class TestVirtualImaging:
     # Tests that virtual imaging runs without failure
 
-    def test_plot_interactive_virtual_image(self, diffraction_profile):
-        diffraction_profile.plot_interactive_virtual_image(left=1., right=2.)
+    def test_plot_interactive_virtual_image(self, electron_diffraction1d):
+        electron_diffraction1d.plot_interactive_virtual_image(left=1., right=2.)
 
-    def test_get_virtual_image(self, diffraction_profile):
-        diffraction_profile.get_virtual_image(left=1., right=2.)
+    def test_get_virtual_image(self, electron_diffraction1d):
+        electron_diffraction1d.get_virtual_image(left=1., right=2.)
