+++ pyxem/tests/test_utils/test_sim_utils.py
@@ -20,7 +20,7 @@ import pytest
 import numpy as np
 import diffpy
 
-from pyxem.signals.electron_diffraction import ElectronDiffraction2D
+from pyxem.signals.electron_diffraction2d import ElectronDiffraction2D
 from pyxem.utils.sim_utils import get_electron_wavelength, \
     get_interaction_constant, get_unique_families, get_kinematical_intensities,\
     get_vectorized_list_for_atomic_scattering_factors, get_points_in_sphere, \
